pyproject.toml
[project]
name = "poc-juma-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "db-dtypes>=1.5.0",
    "google-auth>=2.45.0",
    "google-cloud-bigquery>=3.39.0",
    "pandas>=2.3.3",
    "pandas-gbq>=0.32.0",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
]
